backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.2
google-cloud-videointelligence==2.11.4
azure-ai-textanalytics==5.3.0
requests==2.31.0
moviepy==1.0.3
yt-dlp==2023.11.16
pydantic==2.5.0
httpx<0.25.0,>=0.24.0
celery==5.3.4
redis==5.0.1
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2